Add decodeUriComponent to data_functions.js
